back-Label color uint

+++ RestAPI101.Backend/RestAPI101.Domain/Entities/Label.cs
@@ -10,7 +10,7 @@ namespace RestAPI101.Domain.Entities
 
         public string Description { get; set; }
 
-        public int Color { get; set; }
+        public uint Color { get; set; }
 
         public string UserLogin { get; }
 
@@ -18,7 +18,7 @@ namespace RestAPI101.Domain.Entities
 
         public List<Todo> Todos { get; }
 
-        public Label(int id, string name, string description, int color, string userLogin)
+        public Label(int id, string name, string description, uint color, string userLogin)
         {
             Id = id;
             Name = name;
@@ -28,12 +28,12 @@ namespace RestAPI101.Domain.Entities
             Todos = new List<Todo>();
         }
 
-        public Label(string name, string? description, int? color)
+        public Label(string name, string? description, uint? color)
         {
             Id = default;
             Name = name;
             Description = description ?? "";
-            Color = color ?? name.GetHashCode();
+            Color = color ?? (uint)name.GetHashCode();
             UserLogin = "";
             Todos = new List<Todo>();
         }
